% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{authors_to_json}
\alias{authors_to_json}
\title{Convert citation authors to JSON}
\usage{
authors_to_json(authors)
}
\arguments{
\item{authors}{A list of Person objects}
}
\value{
A list containing each author, with `given` and `family` name fields
}
\description{
Convert citation authors to JSON
}
