% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{citation_to_json}
\alias{citation_to_json}
\title{Convert citation to JSON}
\usage{
citation_to_json(citation)
}
\arguments{
\item{citation}{The citation of a publication, as a `RefManageR::BibEntry`
object}
}
\value{
A JSON representation of the citation
}
\description{
Convert citation to JSON
}
