% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{covariate_definitions_to_json}
\alias{covariate_definitions_to_json}
\title{Convert covariate definitions to JSON}
\usage{
covariate_definitions_to_json(covt_def_data)
}
\arguments{
\item{covt_def_data}{A list of covariate definitions}
}
\value{
A list of covariate definitions parsable to JSON
}
\description{
Convert covariate definitions to JSON
}
