% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo.R
\name{delete_pub}
\alias{delete_pub}
\title{Deletes a publication and its member models from the database}
\usage{
delete_pub(pub_con, model_con, pub_id, verbose = TRUE)
}
\description{
Deletes a publication and its member models from the database
}
