% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo.R
\name{delete_pubs}
\alias{delete_pubs}
\title{Delete a set of publications and their models}
\usage{
delete_pubs(pub_con, model_con, pub_ids, verbose = TRUE)
}
\description{
Delete a set of publications and their models
}
