% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{descriptors_to_json}
\alias{descriptors_to_json}
\title{Convert descriptors to JSON}
\usage{
descriptors_to_json(descriptors)
}
\arguments{
\item{descriptors}{A list containing descriptors}
}
\value{
A list containing descriptors converted to JSON format
}
\description{
Convert descriptors to JSON
}
