% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_processing.R
\name{get_current_pub_ids}
\alias{get_current_pub_ids}
\title{Get the currently stored publication file IDs from the publications directory}
\usage{
get_current_pub_ids()
}
\description{
Get the currently stored publication file IDs from the publications directory
}
