% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_processing.R
\name{get_current_pub_ids_params}
\alias{get_current_pub_ids_params}
\title{Get the currently stored publication file IDs from the parameters directory}
\usage{
get_current_pub_ids_params()
}
\description{
Get the currently stored publication file IDs from the parameters directory
}
