% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{get_model_con}
\alias{get_model_con}
\title{Establish a connection to the models collection}
\usage{
get_model_con(url)
}
\arguments{
\item{url}{A mongodb url connection string}
}
\description{
Establish a connection to the models collection
}
