% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_mod.R
\name{get_modified_files}
\alias{get_modified_files}
\title{Get publication IDs modified since a specified commit}
\usage{
get_modified_files(last_commit)
}
\arguments{
\item{last_commit}{The commit ID to check to}
}
\description{
Get publication IDs modified since a specified commit
}
