% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{get_pub_con}
\alias{get_pub_con}
\title{Establish a connection to the publications collection}
\usage{
get_pub_con(url)
}
\arguments{
\item{url}{A mongodb url connection string}
}
\description{
Establish a connection to the publications collection
}
