% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{get_update_con}
\alias{get_update_con}
\title{Establish a connection to the update collection}
\usage{
get_update_con(url)
}
\arguments{
\item{url}{A mongodb url connection string}
}
\description{
Establish a connection to the update collection
}
