% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\docType{data}
\name{glmnet_2}
\alias{glmnet_2}
\title{Dengue Models}
\format{An object of class \code{train} (inherits from \code{train.formula}) of length 24.}
\usage{
glmnet_2
}
\description{
This package is for storage and recall of models for dengue project.
}
\keyword{forest,}
\keyword{glmnet,}
\keyword{random}
\keyword{rf}
