% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{load_parameter_frame}
\alias{load_parameter_frame}
\title{Load a parameter frame from the models/parameters directory}
\usage{
load_parameter_frame(name)
}
\arguments{
\item{name}{The name of the file, excluding the extension}
}
\value{
A tibble::tbl_df of the parameter data.
}
\description{
This is a convenience function that allows a user to easily load parameter
files from a directory, and is typically used when creating
`FixedEffectsSet`. By default the function will load parameter frames
from an existing `allometric` installation. For the purposes of testing
publication files locally, refer to `set_params_path`.
}
