% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{model_to_json}
\alias{model_to_json}
\title{Convert model to JSON}
\usage{
model_to_json(model)
}
\arguments{
\item{model}{A model object, i.e., `allometric::FixedEffectsModel()`}
}
\value{
A list containing a parsable form of the model object
}
\description{
Convert model to JSON
}
