% Generated by roxygen2 (4.0.2): do not edit by hand
\name{named_list}
\alias{named_list}
\title{named_list}
\usage{
named_list(l)
}
\arguments{
\item{l}{a list.}
}
\value{
a list, with indices as names.
}
\description{
Return a list with list indices as names.
}
\details{
Useful for creating lists that are inputs to l_ply family of
functions, because they save the names in the .id column.
}
\author{
Alexander Jerneck
}

