% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{parse_func_body}
\alias{parse_func_body}
\title{Parse a function into a string}
\usage{
parse_func_body(func_body)
}
\arguments{
\item{func_body}{The body of a function}
}
\value{
A string representation of the function
}
\description{
Parse a function into a string
}
