% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_mod.R
\name{parse_parameter_names}
\alias{parse_parameter_names}
\title{Parse parameter file names for publication IDs}
\usage{
parse_parameter_names(filenames)
}
\arguments{
\item{filenames}{A vector of parameter CSV filenames (including extension)}
}
\value{
A vector of unique publication IDs
}
\description{
Parse parameter file names for publication IDs
}
