% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{prepare_inline_citation}
\alias{prepare_inline_citation}
\title{Generate an inline citation}
\usage{
prepare_inline_citation(citation)
}
\arguments{
\item{citation}{A `RefManageR::BibEntry` object}
}
\value{
A string containing the inline citation
}
\description{
Generate an inline citation
}
