% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{publication_to_json}
\alias{publication_to_json}
\title{Convert a publication to a JSON representation}
\usage{
publication_to_json(publication)
}
\arguments{
\item{publication}{The `allometric::Publication` object}
}
\value{
A list containing fields `pub_json`, which is sent to the
 publications collection and `models_json`, which is a list of JSON
 representations of models in the publication.
}
\description{
This is the main function for processing publications and the models they
contain into JSON, and is used directly by `map_publications`.
}
