% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na.R
\name{replace_na}
\alias{replace_na}
\title{Replace NA from col type is character}
\usage{
replace_na(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame with NA's
}
\description{
Replace NA from col type is character
}
\examples{
data <- data.frame(
a = c("hello", "NA", "", "world"),
b = c(1, 2, 3, 4),
c = c("foo", "bar", "", "NA"),
stringsAsFactors = FALSE)
replace_na_strings(data)
}
