% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run.glm}
\alias{run.glm}
\title{Run a generalized linear model on the formula.}
\usage{
run.glm(formula, ...)
}
\arguments{
\item{formula}{a formula object specifying the model.}

\item{...}{additional arguments passed on to glm.}
}
\description{
This wrapper function is required to get correct model objects
when using lapply.
}

