% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run.lm}
\alias{run.lm}
\title{run.lm}
\usage{
run.lm(formula, ...)
}
\arguments{
\item{formula}{a formula object specifying the model.}

\item{...}{additional arguments passed on to the lm.}
}
\value{
a model object
}
\description{
Run a linear model on the formula.
}
\details{
This wrapper function is required to get correct model objects
when using lapply.
}
\author{
Alexander Jerneck
}

