% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run.nb}
\alias{run.nb}
\title{run.nb}
\usage{
run.nb(formula, ...)
}
\arguments{
\item{formula}{a formula object specifying the model.}

\item{...}{additional arguments passed on to the nb.}
}
\value{
a model object
}
\description{
Run a negative binomial model on the formula.
}
\details{
This wrapper function is required to get correct model objects
when using lapply.
}
\author{
Alexander Jerneck
}

