% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{taxa_to_json}
\alias{taxa_to_json}
\title{Convert a taxa object to JSON}
\usage{
taxa_to_json(taxa)
}
\arguments{
\item{taxa}{An `allometric::Taxa()` object}
}
\value{
A list of taxons in list format
}
\description{
Convert a taxa object to JSON
}
