% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{unbox_nested}
\alias{unbox_nested}
\title{Unbox the elements of a nested list}
\usage{
unbox_nested(object)
}
\arguments{
\item{object}{A nested list}
}
\value{
A list with nested elements unboxed
}
\description{
Unbox the elements of a nested list
}
