% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{unbox_nonnested}
\alias{unbox_nonnested}
\title{Unbox the elements of a list}
\usage{
unbox_nonnested(object)
}
\arguments{
\item{object}{A nonnested list}
}
\value{
A list with each element unboxed
}
\description{
Unbox the elements of a list
}
