% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_categorical.R
\name{univariate_categorical}
\alias{univariate_categorical}
\title{univariate_categorical, grafic explore to categorical variable}
\usage{
univariate_categorical(
  data,
  feature,
  target = "target",
  labels = c("No", "Yes"),
  colors = c("#07bbc3", "#fb756e"),
  ylog = FALSE,
  label_rotation = FALSE,
  horizontal_layout = TRUE,
  bar_desc = FALSE
)
}
\arguments{
\item{data}{a data frame}

\item{feature}{name from variable}

\item{target}{name from target}

\item{labels}{categories that take target}

\item{colors}{colors that take target}

\item{ylog}{scale log the feature variable}

\item{label_rotation}{rotation the y label}

\item{horizontal_layout}{show horizontal or vertical}

\item{bar_desc}{order bar plot}
}
\value{
a plot from explore categorical variable
}
\description{
univariate_categorical, grafic explore to categorical variable
}
\examples{
example_data <- data.frame(
  Sex = c("Male", "Female", "Male", "Female", "Male"),
  Survived = c(0, 1, 0, 1, 0)
)
univariate_categorical(data = example_data, feature = "Sex", target = "Survived")
}
