% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo.R
\name{update_db}
\alias{update_db}
\title{Update the database on push}
\usage{
update_db(pub_con, model_con, update_con, verbose = TRUE, last_commit = NULL)
}
\description{
This is the main function for updating the database. It proceeds in five
phases, meant to remove old/"hangnail" publications and models.
 1. Delete all nonexistent publications from db (i.e., in publications
    no longer in directory) and their models
 2. Determine which publications have been modified since last commit
 3. Generate JSON for modified publications+models, retain newly generated
    model IDs
 4. Upsert the JSON for modified publications
 5. Delete any models belonging to the modified publications that are not in
    the set of new model IDs. These should represent modified models
}
