% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo.R
\name{upsert_publication}
\alias{upsert_publication}
\title{Update a publication and its models in mongodb}
\usage{
upsert_publication(pub_con, model_con, pub_json, models_json, verbose = TRUE)
}
\arguments{
\item{pub_con}{A mongodatabase connection to the publications collection}

\item{model_con}{A mongodatabase connection to the models collection}
}
\description{
This is the main function for updating the DB. This is done by upserting
the publication (which contains model references) and then upserting each
individual model
}
