% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{variables_to_json}
\alias{variables_to_json}
\title{Convert variables to JSON}
\usage{
variables_to_json(variables)
}
\arguments{
\item{variables}{A list containing variables, via `response` or `covariates`
slots of a model}
}
\value{
A list of parsed variables
}
\description{
Convert variables to JSON
}
